methode Tabou

+++ cvrp/src/tests/cvrp.rs
@@ -5,7 +5,7 @@ mod cvrp {
 
     #[test_case(200, 100; "When mod is 0")]
     #[test_case(200, 110; "When mod is not 0")]
-    fn test_max_trucks(max_weight: i32, max_truck_weight: i32) {
+    fn test_max_trucks(max_weight: u16, max_truck_weight: u16) {
         let cvrp = CVRP::mock(None, None, Some(max_truck_weight), Some(max_weight));
         assert_eq!(2, cvrp.get_max_nb_truck())
     }
